Get more tests into general layer
