+++ stores/_test-utils/src/default-tests.ts
@@ -12,9 +12,11 @@ import { createScoresTest } from './domains/scores';
 import { createConversationsTest } from './domains/conversations';
 import { createTestSuiteWorkflows } from './domains/workflows';
 import { createTraceTests } from './domains/traces';
+import { createEvalsTests } from './domains/evals';
 
 export * from './domains/conversations/data';
 export * from './domains/workflows/data';
+export * from './domains/evals/data';
 
 export const createSampleTraceForDB = (
   name: string,
@@ -38,23 +40,7 @@ export const createSampleTraceForDB = (
   createdAt: createdAt || new Date(),
 });
 
-export const createSampleEval = (agentName: string, isTest = false, createdAt?: Date) => {
-  const testInfo = isTest ? { testPath: 'test/path.ts', testName: 'Test Name' } : undefined;
 
-  return {
-    agent_name: agentName,
-    input: 'Sample input',
-    output: 'Sample output',
-    result: { score: 0.8 } as MetricResult,
-    metric_name: 'sample-metric',
-    instructions: 'Sample instructions',
-    test_info: testInfo,
-    global_run_id: `global-${randomUUID()}`,
-    run_id: `run-${randomUUID()}`,
-    created_at: createdAt || new Date().toISOString(),
-    createdAt: createdAt || new Date(),
-  };
-};
 
 export function createTestSuite(storage: MastraStorage) {
   describe(storage.constructor.name, () => {
@@ -82,6 +68,8 @@ export function createTestSuite(storage: MastraStorage) {
     createTestSuiteWorkflows(storage);
 
     createTraceTests(storage);
+
+    createEvalsTests(storage);
 
 
   });
@@ -129,106 +117,7 @@ export function createTestSuite(storage: MastraStorage) {
 
   });
 
-  describe('Eval Operations', () => {
-    const createSampleEval = (agentName: string, isTest = false) => {
-      const testInfo = isTest ? { testPath: 'test/path.ts', testName: 'Test Name' } : undefined;
 
-      return {
-        id: randomUUID(),
-        agentName,
-        input: 'Sample input',
-        output: 'Sample output',
-        result: { score: 0.8 } as MetricResult,
-        metricName: 'sample-metric',
-        instructions: 'Sample instructions',
-        testInfo,
-        globalRunId: `global-${randomUUID()}`,
-        runId: `run-${randomUUID()}`,
-        createdAt: new Date().toISOString(),
-      };
-    };
-
-    it('should retrieve evals by agent name', async () => {
-      const agentName = `test-agent-${randomUUID()}`;
-
-      // Create sample evals
-      const liveEval = createSampleEval(agentName, false);
-      const testEval = createSampleEval(agentName, true);
-      const otherAgentEval = createSampleEval(`other-agent-${randomUUID()}`, false);
-
-      // Insert evals
-      await storage.insert({
-        tableName: TABLE_EVALS,
-        record: {
-          agent_name: liveEval.agentName,
-          input: liveEval.input,
-          output: liveEval.output,
-          result: liveEval.result,
-          metric_name: liveEval.metricName,
-          instructions: liveEval.instructions,
-          test_info: null,
-          global_run_id: liveEval.globalRunId,
-          run_id: liveEval.runId,
-          created_at: liveEval.createdAt,
-          createdAt: new Date(liveEval.createdAt),
-        },
-      });
-
-      await storage.insert({
-        tableName: TABLE_EVALS,
-        record: {
-          agent_name: testEval.agentName,
-          input: testEval.input,
-          output: testEval.output,
-          result: testEval.result,
-          metric_name: testEval.metricName,
-          instructions: testEval.instructions,
-          test_info: JSON.stringify(testEval.testInfo),
-          global_run_id: testEval.globalRunId,
-          run_id: testEval.runId,
-          created_at: testEval.createdAt,
-          createdAt: new Date(testEval.createdAt),
-        },
-      });
-
-      await storage.insert({
-        tableName: TABLE_EVALS,
-        record: {
-          agent_name: otherAgentEval.agentName,
-          input: otherAgentEval.input,
-          output: otherAgentEval.output,
-          result: otherAgentEval.result,
-          metric_name: otherAgentEval.metricName,
-          instructions: otherAgentEval.instructions,
-          test_info: null,
-          global_run_id: otherAgentEval.globalRunId,
-          run_id: otherAgentEval.runId,
-          created_at: otherAgentEval.createdAt,
-          createdAt: new Date(otherAgentEval.createdAt),
-        },
-      });
-
-      // Test getting all evals for the agent
-      const allEvals = await storage.getEvalsByAgentName(agentName);
-      expect(allEvals).toHaveLength(2);
-      expect(allEvals.map(e => e.runId)).toEqual(expect.arrayContaining([liveEval.runId, testEval.runId]));
-
-      // Test getting only live evals
-      const liveEvals = await storage.getEvalsByAgentName(agentName, 'live');
-      expect(liveEvals).toHaveLength(1);
-      expect(liveEvals?.[0]?.runId).toBe(liveEval.runId);
-
-      // Test getting only test evals
-      const testEvals = await storage.getEvalsByAgentName(agentName, 'test');
-      expect(testEvals).toHaveLength(1);
-      expect(testEvals?.[0]?.runId).toBe(testEval.runId);
-      expect(testEvals?.[0]?.testInfo).toEqual(testEval.testInfo);
-
-      // Test getting evals for non-existent agent
-      const nonExistentEvals = await storage.getEvalsByAgentName('non-existent-agent');
-      expect(nonExistentEvals).toHaveLength(0);
-    });
-  });
 
   describe('alterTable', () => {
     const TEST_TABLE = 'test_alter_table';
